+++ package.json
@@ -35,7 +35,9 @@
   },
   "packageManager": "yarn@4.9.4+sha512.7b1cb0b62abba6a537b3a2ce00811a843bea02bcf53138581a6ae5b1bf563f734872bd47de49ce32a9ca9dcaff995aa789577ffb16811da7c603dcf69e73750b",
   "dependencies": {
+    "@mdi/font": "^7.4.47",
     "comlink": "^4.4.2",
+    "pinia": "^3.0.3",
     "plotly.js": "3.1.0",
     "sprintf-js": "^1.1.3",
     "vue": "^3.5.21",
