+++ src/main.ts
@@ -2,5 +2,6 @@ import { createApp } from 'vue';
 
 import App from './App.vue';
 import vuetify from './plugins/vuetify';
+import pinia from './store';
 
-createApp(App).use(vuetify).mount('#app');
+createApp(App).use(vuetify).use(pinia).mount('#app');
