+++ src/plugins/vuetify.ts
@@ -1,5 +1,6 @@
 // Vuetify
 import 'vuetify/styles';
+import '@mdi/font/css/materialdesignicons.css';
 
 import { createVuetify } from 'vuetify';
 import * as components from 'vuetify/components';
