+++ src/store/index.ts
@@ -1,0 +1,3 @@
+import { createPinia } from 'pinia';
+
+export default createPinia();
